requirements.txt
discord.py>=2.0.0
py-cord
google-generativeai
python-dotenv